+++ Autoloads/Utils.gd
@@ -1,6 +1,0 @@
-class_name Utils
-
-const MIN_SIZE_X: float = 96.0
-const MIN_SIZE_Y: float = 96.0
-const MAX_SIZE_X: float = 96.0
-const MAX_SIZE_Y: float = 96.0